chore(project): change provider to store
* review fixes
