+++ src/App.tsx
@@ -1,15 +1,18 @@
 import React, { Component } from 'react';
 import { getI18n, I18nextProvider } from 'react-i18next';
 
+import { DEFAULT_CONFIG_LOCATION } from './config';
+
 import type { Config } from '#types/Config';
 import Router from '#src/components/Router/Router';
 import Root from '#src/components/Root/Root';
-import ConfigProvider from '#src/providers/ConfigProvider';
+import LoadingOverlay from '#src/components/LoadingOverlay/LoadingOverlay';
 import QueryProvider from '#src/providers/QueryProvider';
 import { restoreWatchHistory } from '#src/stores/WatchHistoryController';
 import { initializeAccount } from '#src/stores/AccountController';
 import { initializeFavorites } from '#src/stores/FavoritesController';
 import { logDev } from '#src/utils/common';
+import { loadAndValidateConfig } from '#src/utils/config';
 import { PersonalShelf } from '#src/enum/PersonalShelf';
 
 import '#src/i18n/config';
@@ -17,11 +20,13 @@ import '#src/styles/main.scss';
 
 interface State {
   error: Error | null;
+  isLoading: boolean;
 }
 
 class App extends Component {
   public state: State = {
     error: null,
+    isLoading: false,
   };
 
   componentDidCatch(error: Error) {
@@ -45,32 +50,43 @@ class App extends Component {
   }
 
   configLoadingHandler = (isLoading: boolean) => {
+    this.setState({ isLoading });
     logDev(`Loading config: ${isLoading}`);
   };
 
   configErrorHandler = (error: Error) => {
     this.setState({ error });
+    this.setState({ isLoading: false });
     logDev('Error while loading the config.json:', error);
   };
 
   configValidationCompletedHandler = async (config: Config) => {
+    this.setState({ isLoading: false });
     await this.initializeServices(config);
   };
 
+  componentDidMount() {
+    loadAndValidateConfig(
+      window.configLocation || DEFAULT_CONFIG_LOCATION,
+      this.configLoadingHandler,
+      this.configErrorHandler,
+      this.configValidationCompletedHandler,
+    );
+  }
+
   render() {
+    const { isLoading, error } = this.state;
+
+    if (isLoading) {
+      return <LoadingOverlay />;
+    }
+
     return (
       <I18nextProvider i18n={getI18n()}>
         <QueryProvider>
-          <ConfigProvider
-            configLocation={window.configLocation || '/config.json'}
-            onLoading={this.configLoadingHandler}
-            onValidationError={this.configErrorHandler}
-            onValidationCompleted={this.configValidationCompletedHandler}
-          >
-            <Router>
-              <Root error={this.state.error} />
-            </Router>
-          </ConfigProvider>
+          <Router>
+            <Root error={error} />
+          </Router>
         </QueryProvider>
       </I18nextProvider>
     );
