+++ src/config.ts
@@ -13,3 +13,5 @@ export const MAX_WATCHLIST_ITEMS_COUNT = 48;
 export const ADYEN_TEST_CLIENT_KEY = 'test_I4OFGUUCEVB5TI222AS3N2Y2LY6PJM3K';
 
 export const ADYEN_LIVE_CLIENT_KEY = 'live_BQDOFBYTGZB3XKF62GBYSLPUJ4YW2TPL';
+
+export const DEFAULT_CONFIG_LOCATION = '/config.json';
